requirements.txt
gunicorn==20.1.0
psycopg2-binary==2.8.6
google-cloud-secret-manager==2.4.0
google-auth==1.30.0
django-storages[google]==1.11.1
django-environ==0.4.5
google-cloud-pubsub==1.7.0
